requirements.txt
Flask_Login==0.4.0
Werkzeug==0.12.2
SQLAlchemy==1.1.14
PyMySql==0.7.11
Flask-WTF==0.14.2
WTForms==2.1
flask_excel == 0.0.7
xlsxwriter == 1.0.0
pyexcel-xls==0.5.0
flask_mail == 0.9.1